backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.2.11
djangorestframework==3.16.1
feedparser==6.0.12
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.3.1
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
